requirements.txt
cycler==0.10.0
matplotlib==2.1.0
numpy==1.13.3
pandas==0.22.0
pkg-resources==0.0.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn-pandas==1.6.0
